backend/fastapi/requirements.txt
# FastAPI backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.0
python-dotenv==1.0.0
pydantic==1.10.11
cryptography==42.0.0
python-multipart==0.0.6

